complete 3 Cases  API

+++ src/main/java/hanghae99/clonecoding/airbnb/repository/HotelRepository.java
@@ -4,6 +4,8 @@ import hanghae99.clonecoding.airbnb.entity.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface HotelRepository extends JpaRepository<Hotel,Integer> {
+public interface HotelRepository extends JpaRepository<Hotel,Long> {
 }
